Убираем старые "мозги" нафиг
